extractAfter removed and image figures started
extractAfter is a function introduced in 2016b.  I've replaced it so
that this code can run on earlier versions.

I changed the location of the parameter directory.

Finally, I started prompting the code to display the figures at the end.
Still need to write function saving those figures as tifs.