Began script to format polscope data
The polscope data cannot be used with the stitching/grid algorithm in
FIJI.  This script is meant to reformat the images/metadata into the
right form.  Currently, it can read in a list of folders and open the
metadata for each folder.

However, the sorting/reformatting logic has not yet been implemeneted.