Jupyter/SimpleITK Registration
I am redoing the registration workflow for Jupyter and SimpeITK.  This
creates an interactive coding environment, where individual parts can be
rerun on command, and where the data is mixed in with the code.

So far, I just have it pick two specific images from one of my
directories and then display them.  Registration/Segmentation and more
to come later.