MMP-Pol registration and CW-SSIM analysis complete
Completed the registration, though I had to use offset origins for
samples WP5 and WP9, as they were translated too far for the optimizer
to converge.

Made a mask over the result, placing it on the fixed image, and wrote
that for quantitative comparisons using SSIM.

Performed SSIM analysis, getting good results (3 ~ 0.8, 1 ~0.66)

Performed colocalization analysis on WP2, but it was not very promising.
Better than before, but not reportable.