Found spacing error - new SHG checkpoint
Found an error with the way ITK saves spacing.  Instead of 5.416 mm, it
was saving as 5.47 - this didn't matter much for the polscope-mmp
registration (as I was calibrating MMP spacing off of the polscope image
manually), but it potentially mattered a lot for the SHG, where I just
took down the known resolution and didn't manually compare.