Function to take a small ROI and calculate the average retardance over it
This function nominally calculates an average retardance value over a
region of interest.

It does this by first doubling the orientation angle (as it starts a
0-180), then calculating the resulting angle vector for each pixel.

This is then multiplied by the retardance matrix, and summed up in an
average.

The resulting retardance magnitude is taken from the absolute value, and
the angle is the phase angle of the complex number divided by 2 (since
the original image goes from 0-180).