Moved calculating the number of tiles to a separate function
I moved the code that calculates the number of tiles into a separate
function, to make it more readable and make it replicable in the future.

This function returns the number of tiles that fits along a particular
image dimension., along with an offset that centers the tiles, leaning
to the left if there is an odd number of remaining pixels.